Add a new argument "--download-only"
The argument provides simple file downloading interface. It doesn't
produce any additional files such as JSON or CSV.
The other arguments for processing downloaded files (such as --unzip)
will still be proceeded.

The main reason to add this option is to give a clener feature of
making the tool a clean sync application. It can be used to sync with
active Dropbox directory which can be a hub data source for other
applications such as Golden Cheetah or Training Peaks.

As this option has a huge impact on the final results of processing
the application, there is no shortcut version of the argument.